add login logic buttons for testing purpose
